requirements.txt
django==4.2.0
Pillow==10.2.0
python-dotenv=1.0.1
psycopg2-binary==2.9.9
djangorestframework==3.14.0
black==24.3.0